==> Figure5/02_TEAseq/01A_Download_data_from_GEO.ipynb <==
# load libraries
library(dplyr)
library(GEOquery)

# Download TEAseq Data

options(timeout = 2000)  
# stop this if file size is 29.9GB
url <- 'ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE214nnn/GSE214546/suppl/GSE214546_RAW.tar'
destfile <- 'GSE214546_RAW.tar'
download.file(url, destfile, method = 'wget')

untar("GSE214546_RAW.tar", exdir = "GSE214546_Data")

# Get MetaData

eList <- getGEO("GSE214546")
df_meta<-as.data.frame(eList$GSE214546_series_matrix.txt.gz) %>% arrange(geo_accession)
df_meta<-df_meta[c('geo_accession','age.ch1',
                   'birth.year.ch1','cell.type.ch1',
                   'library.type.ch1','race.ch1',
                   'Sex.ch1','subject_id.ch1')]
colnames(df_meta)[1]<-'GEO_Accession'

##only tea seq library has h5 with frag file
all_h5 <- list.files(path = 'GSE214546_Data/', pattern = '*frag*', full.names = TRUE)
all_h5<-all_h5[1:16]
extracted_pattern <- sub(".*/(.*?)_[^_]*$", "\\1", all_h5)

split_list <- strsplit(extracted_pattern, "_", fixed = TRUE)
split_df <- do.call(rbind, sapply(split_list, function(x) as.data.frame(t(x)), simplify = FALSE))
df <- as.data.frame(split_df, stringsAsFactors = FALSE)

colnames(df)<-c('GEO_Accession','Batch','pbmc_sample_id')
df$combined_sample_id<-extracted_pattern

df<-left_join(df,df_meta)

write.csv(df,'meta_data_GEO.csv')
