==> Figure5/02_TEAseq/03A_ATAC_Processing.ipynb <==
quiet_library <- function(...) {
    suppressPackageStartupMessages(library(...))
}
quiet_library('tidyverse')
quiet_library("hise")
quiet_library('ArchR')
quiet_library('data.table')
quiet_library('parallel')
quiet_library('arrow')
quiet_library('dplyr')
library(BSgenome.Hsapiens.UCSC.hg38)

options(future.globals.maxSize = 1000 * 1024^5)

addArchRGenome("hg38")
addArchRThreads(threads = 10)
pathToMacs2<-'/opt/conda/bin/macs3'

sessionInfo()

# we might need the specific version of these two packages to make sure there is no bugs :)
#remotes::install_version("ggplot2", version = "3.3.0")
#remotes::install_version("Matrix", version = "1.6.3")

# Create Arrow file and ArchR project

meta<-read.csv('meta_data_GEO.csv')

for (i in meta$combined_sample_id){
ArrowFiles <- createArrowFiles(
  inputFiles = paste0("GSE214546_Data/",i,'_fragments.tsv.gz'),
  sampleNames = i,
  minTSS = 4, #Dont set this too high because you can always increase later
  minFrags = 1000, 
  addTileMat = TRUE,
  addGeneScoreMat = TRUE
)}

projHeme1 <- ArchRProject(
  ArrowFiles = paste0(meta$combined_sample_id,'.arrow'), 
  outputDirectory = "PenSen_ATAC",
  copyArrows = TRUE)

# Keep Cells that with RNA module 

# Read scRNA data
adata_scRNA <- anndata::read_h5ad('TEA_scRNA.h5ad', backed='r')

# Extract and process scRNA labels
scRNA_labels <- adata_scRNA$obs
scRNA_labels$real_barcodes <- paste0(scRNA_labels$well_id, '-', scRNA_labels$original_barcodes, '-1')

# Extract and process scATAC labels
scATAC_labels <- as.data.frame(projHeme1@cellColData)
scATAC_labels$barcodes <- str_extract(rownames(scATAC_labels), "(?<=#).*$")

# Generate metadata file paths
scATAC_meta_data_files <- paste0("GSE214546_Data/", meta$combined_sample_id, '_filtered_metadata.csv.gz')

# Read and combine metadata for scATAC
meta_data_list_ATAC <- mclapply(scATAC_meta_data_files, function(x) {
  metadata <- read.csv(gzfile(x))
  return(metadata)
}, mc.cores = 16)

meta_data_ATAC <- do.call(rbind, meta_data_list_ATAC)

# Merge scATAC labels with metadata
scATAC_labels <- left_join(scATAC_labels, meta_data_ATAC, by = 'barcodes')

# Create real barcodes for scATAC labels
scATAC_labels$real_barcodes <- paste0(gsub('-A', '-', scATAC_labels$well_id), '-', scATAC_labels$original_barcodes)

# Merge scATAC labels with scRNA labels
scATAC_labels <- left_join(scATAC_labels, scRNA_labels, by = 'real_barcodes')

# Extract barcodes for projHeme1
projHeme1@cellColData$barcodes <- str_extract(rownames(as.data.frame(projHeme1@cellColData)), "(?<=#).*$")

# Create ATAC barcodes for scATAC labels
scATAC_labels$ATAC_barcodes <- paste0(scATAC_labels$Sample, '#', scATAC_labels$barcodes.x)

# Filter scATAC labels to make sure it is pure enough
scATAC_labels <- scATAC_labels %>%
  filter(AIFI_L1 == 'T cell') %>%
  filter(AIFI_L2 %in% c('Treg', 'Proliferating T cell', 'Naive CD8 T cell', 'Naive CD4 T cell', 
                        'CD8aa', 'MAIT', 'Memory CD8 T cell', 'Memory CD4 T cell', 'DN T cell'))

# Subset projHeme1 based on filtered scATAC labels
projHeme1 <- projHeme1[scATAC_labels$ATAC_barcodes,]

# Update projHeme1 cellColData with labels
projHeme1@cellColData$AIFI_L1 <- scATAC_labels$AIFI_L1
projHeme1@cellColData$AIFI_L2 <- scATAC_labels$AIFI_L2
projHeme1@cellColData$AIFI_L3 <- scATAC_labels$AIFI_L3

# Save it
saveArchRProject(ArchRProj = projHeme1, outputDirectory = "PenSen_ATAC", load = TRUE)

# Processing

# Load the ArchR project
projHeme1 <- loadArchRProject(path = 'PenSen_ATAC/')

# Add iterative LSI for dimensionality reduction(might get an error due to ggplot2 version issue, but it is safe to proceed)
projHeme1 <- addIterativeLSI(projHeme1, name = 'IterativeLSI', force = TRUE, varFeatures = 75000)

# Compute group coverages by "AIFI_L3", using up to 2000 cells and 16 replicates per group
projHeme1 <- addGroupCoverages(ArchRProj = projHeme1, 
                               groupBy = "AIFI_L3", 
                               maxReplicates = 16, 
                               maxCells = 2000, 
                               threads = 10, 
                               force = TRUE)

# Identify reproducible peaks using MACS2, grouped by "AIFI_L3"
projHeme1 <- addReproduciblePeakSet(ArchRProj = projHeme1, 
                                    groupBy = "AIFI_L3", 
                                    pathToMacs2 = pathToMacs2, 
                                    force = TRUE)

# Add a peak matrix
projHeme1 <- addPeakMatrix(projHeme1, force = TRUE)

# Annotate peaks with motifs from the "cisbp" set
projHeme1 <- addMotifAnnotations(ArchRProj = projHeme1, 
                                 motifSet = "cisbp", 
                                 name = "cisbp", force = TRUE)

# Add background peaks
projHeme1 <- addBgdPeaks(projHeme1, force = TRUE)

# Add a deviations matrix for "Motif" peaks
projHeme1 <- addDeviationsMatrix(ArchRProj = projHeme1, 
                                 peakAnnotation = "Motif", 
                                 force = TRUE, 
                                 threads = 50)

# Add impute weights
projHeme1 <- addImputeWeights(projHeme1,seed=1)

# Save it
saveArchRProject(ArchRProj = projHeme1, outputDirectory = "PenSen_ATAC", load = TRUE)


projHeme1 <- loadArchRProject(path = 'PenSen_ATAC/')


MotifMatrix<-getMatrixFromProject(
  ArchRProj = projHeme1,
  useMatrix = "MotifMatrix",
  useSeqnames = NULL,
  verbose = TRUE,
  binarize = FALSE,
  threads = getArchRThreads(),
  logFile = createLogFile("getMatrixFromProject")
)

ChromVar_Z_Imputed<-imputeMatrix(
  mat = MotifMatrix@assays@data$z,
  imputeWeights =  getImputeWeights(projHeme1),
  threads = 30,
  verbose = TRUE)


# Calculate Mean Mofit score per celltype on each sample

df<-as.data.frame(as.matrix(ChromVar_Z_Imputed))
df$row_names <- paste0("motif_",rownames(df))

df<-t(df)
colnames(df)<-paste0("motif_",colnames(df))
df<-df[1:(dim(df)[1]-1),]

cell_meta_data<-as.data.frame(projHeme1@cellColData)
cell_meta_data<-cell_meta_data[rownames(df),]

table(rownames(cell_meta_data)==rownames(df))

df_combined<-cbind(cell_meta_data,df)

write.csv(df_combined,'motif_score_all_cell.csv')

motif_mean<-df_combined%>%
  group_by(Sample, AIFI_L3) %>%
  dplyr::summarise(across(contains("motif_"), ~ mean(as.numeric(.), na.rm = TRUE), 
                          .names = "{col}"), .groups = 'drop')

motif_mean<-motif_mean %>% ungroup %>% as.data.frame()

write.csv(motif_mean,"Motif_Mean_Score.csv")

sessionInfo()
